chore: refactor trimesh raycast into its own file
