+++ src/query/ray/mod.rs
@@ -28,5 +28,7 @@ mod ray_round_shape;
 mod ray_support_map;
 mod ray_triangle;
 #[cfg(feature = "alloc")]
+mod ray_trimesh;
+#[cfg(feature = "alloc")]
 mod ray_voxels;
 mod simd_ray;
